requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
pyperclip>=1.8.2
asyncio 